stringdist_api {stringdist}	R Documentation

Calling stringdist from C or C++

Description

As of version 0.9.5.0 several C level functions can be linked to and called from C code in other R packages.

A description of the API can be found in stringdist_api.pdf.

[Package stringdist version 0.9.5.1 Index]